1-byte mode checks for byte minimum value (0) rather than short minimum value

+++ BrawlLib/System/Windows/Forms/Section Editor/SectionEditor.cs
@@ -1110,7 +1110,7 @@ namespace System.Windows.Forms
                         hexBox1.Invalidate();
                         PosChanged();
                     }
-                    else if (l < short.MinValue)
+                    else if (l < byte.MinValue)
                     {
                         tempB = byte.MinValue;
                         long t = Position;
